move DevicToken into Auth
