+++ src/Controllers/LoginController.php
@@ -15,6 +15,8 @@ namespace Elabftw\Controllers;
 use Elabftw\Auth\AnonymousLoginContext;
 use Elabftw\Auth\AnonymousLoginValidator;
 use Elabftw\Auth\Demo;
+use Elabftw\Auth\DeviceToken;
+use Elabftw\Auth\DeviceTokenValidator;
 use Elabftw\Auth\External;
 use Elabftw\Auth\InitialTeamSelectionRequired;
 use Elabftw\Auth\Ldap;
@@ -54,8 +56,6 @@ use Elabftw\Models\Users\ExistingUser;
 use Elabftw\Models\Idps;
 use Elabftw\Models\Users\Users;
 use Elabftw\Models\Users2Teams;
-use Elabftw\Services\DeviceToken;
-use Elabftw\Services\DeviceTokenValidator;
 use Elabftw\Services\Filter;
 use Elabftw\Services\TeamsHelper;
 use Elabftw\Services\ResetPasswordKey;
